limiting "dangerous" endpoints (delete operations, most updates, most creates) to admin users only.
Admins are determined by the `pluto.admin-groups` configuration setting, if a logged in user is in any of these AD groups then they are considered an admin.

Checks are disabled in CI mode otherwise it screws up a lot of the tests, I can't make the mocked cache API inject the right values at the moment.  There is therefore a warning in case an attacker somehow manages to corrupt the production environment, to tell the (real) admin that CI mode is active